Patch contributed by Stefan Reinauer: + Compile fixes for GCC 3.3. + Graphical menu support (not enabled yet).
git-svn-id: https://svn.aros.org/svn/aros/trunk@18105 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/i386-pc/boot/grub/stage2/fsys_reiserfs.c
@@ -112,7 +112,7 @@ struct reiserfs_journal_header {
   /* offset in the log of where to start replay after a crash */
   __u32 j_first_unflushed_offset;
   /* mount id to detect very old transactions */
-  __u32 long j_mount_id;
+  __u32 j_mount_id;
 };
 
 /* magic string to find desc blocks in the journal */
